1. Any expression of the right type will do: Wherever a value of a type is needed one can provide a literal value, a variable, or a function application expression of that type. The expression is evaluated to produce a value, which is used in the computation. For example, you can use a literal, variable, or a more complex expression on the right hand side of a function evaluation rule. Here's an example: Re-implement the bool_nand function with just one rule: bool_nand b1 b2 = bool_not (bool_and b1 b2).
2. Conditionals. We define a new function, a ternary function, that returns
either the second or the third of its Boolean arguments depending on the value
of the first argument. The first argument is called the condition. The second
is the "true branch." The third is the "false branch." If the condition is true
the function returns the value of the true branch, otherwise it returns the
value of the false branch.

3. Boolean algebra in the real world: Google search. How do I find pages that
contain "Poodle" or "Pomeranian" but not necessarily both and not "Wikipedia"?